config.sh: create x86_64 Ubuntu host to verify setup
Create  an empty VirtualBox x86_64 VM with the MAC address as specified in host_0_mac.

config.sh will create the host in Foreman. This simple setup helps to verify that the network settings and the  Foreman configuration and Ubuntu VM works in general.

If the network boot works, you can move on to more comlicated configurations.